+++ Network/includes/Protocol/SuccessConnect.hpp
@@ -13,7 +13,7 @@
 
 class SuccessConnect : public Message {
 public:
-    SuccessConnect(uint8_t id)
+    SuccessConnect(size_t id)
             : Message(TCP_COMMUNICATION::SUCCESS_CONNECT)
             , m_idPlayer(id) {}
     ~SuccessConnect() override = default;
@@ -23,9 +23,9 @@ public:
     [[nodiscard]] std::vector<uint8_t> serialize() const override;
     size_t getSize() const final { return PACKET_HDR_SIZE + SUCCESS_CONNECT_SIZE; }
 
-    uint8_t getId() const { return m_idPlayer; }
+    size_t getId() const { return m_idPlayer; }
 private:
-    uint8_t m_idPlayer;
+    size_t m_idPlayer;
 };
